Moved the image agent onto the cache the app already configures
The added cache manager sat outside configureImageDiskCache, so its files escaped the disk budget, clear cache left them behind, and Apple TV lost its JSON repo. Setting the file service on the existing config covers every cached image with no call site changes. The widgets that fetch through Flutter's own image client pass the agent as a header instead.
