+++ lib/data/repositories/media_bar_repository.dart
@@ -8,7 +8,6 @@ import 'package:get_it/get_it.dart';
 import 'package:server_core/server_core.dart';
 
 import '../../preference/user_preferences.dart';
-import '../../util/server_image_cache.dart';
 import '../models/media_bar_slide_item.dart';
 import '../models/media_bar_state.dart';
 
@@ -300,24 +299,12 @@ class MediaBarRepository {
   void precacheImages(BuildContext context, List<MediaBarSlideItem> items) {
     for (final item in items.take(_precacheBackdropCount)) {
       if (item.backdropUrl != null) {
-        precacheImage(
-          CachedNetworkImageProvider(
-            item.backdropUrl!,
-            cacheManager: serverImageCacheManager,
-          ),
-          context,
-        );
+        precacheImage(CachedNetworkImageProvider(item.backdropUrl!), context);
       }
     }
     for (final item in items.take(_precacheLogoCount)) {
       if (item.logoUrl != null) {
-        precacheImage(
-          CachedNetworkImageProvider(
-            item.logoUrl!,
-            cacheManager: serverImageCacheManager,
-          ),
-          context,
-        );
+        precacheImage(CachedNetworkImageProvider(item.logoUrl!), context);
       }
     }
   }
