+++ lib/ui/widgets/bounded_network_image.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
-import '../../util/server_image_cache.dart';
 import '../theme/vibrance.dart';
 import 'image_source.dart';
 
@@ -82,10 +81,7 @@ class BoundedNetworkImage extends StatelessWidget {
       ResizeImage.resizeIfNeeded(
         cacheW,
         null,
-        CachedNetworkImageProvider(
-          imageUrl,
-          cacheManager: serverImageCacheManager,
-        ),
+        CachedNetworkImageProvider(imageUrl),
       ),
       context,
     );
@@ -132,7 +128,6 @@ class BoundedNetworkImage extends StatelessWidget {
         }
         return CachedNetworkImage(
           imageUrl: imageUrl,
-          cacheManager: serverImageCacheManager,
           fit: fit,
           alignment: alignment,
           fadeInDuration: fadeInDuration,
